Added func with gl_FragCoord to particle frag. Fixed to invert normals if facing back faces for lighting in model_loading and added Interface block

+++ LearnOpenGL/src/particle_effects.h
@@ -43,7 +43,7 @@ public:
 		// Bind skybox and textures
 		glBindVertexArray(mVAO);
 		shader.setVec3("colour", mColour);
-		shader.setInt("pointsCount", mParticleCoordinates.size());
+		shader.setInt("pointsCount", (int)mParticleCoordinates.size());
 		// Draw skybox
 		glDrawArrays(GL_POINTS, 0, 3 * (GLsizei)mParticleCoordinates.size());
